1，订单和订单详情

+++ src/main/java/com/kz/core/dao/BaseMapper.java
@@ -7,17 +7,17 @@ import com.kz.po.User;
 
 public interface BaseMapper<M, QM> {
 	
-	public int deleteByPrimaryKey(Serializable uuid);
+	public Long deleteByPrimaryKey(Serializable uuid);
 
-	public int insert(M m);
+	public Long insert(M m);
 
-	public int insertSelective(M m);
+	public Long insertSelective(M m);
 
 	public M selectByPrimaryKey(Serializable uuid);
 
-	public int updateByPrimaryKeySelective(M m);
+	public Long updateByPrimaryKeySelective(M m);
 
-	public int updateByPrimaryKey(M m);
+	public Long updateByPrimaryKey(M m);
 
 	public  List<M> getByConditionPage(QM qm);
 }
